requirements.txt
# Dependencies for Speech-to-Text API
google-genai
fastapi
uvicorn[standard]
python-multipart
python-dotenv
requests  # For testing

# TTS Dependencies
transformers
torch
scipy
